test(cua-computer): split parity fixtures under the line cap and fix strict test types
